test: relocate easy-win tests to subject-matched files
Move the tests added in 9343282 into the existing per-subject files
matching the repo convention (one class per subject area):

- MonsterAllocateStats / HP scaling -> Hybrasyl.Tests/Monster.cs
- ItemObject FormulaVariable annotations -> Hybrasyl.Tests/Items.cs
- Cone radius cap regression -> Hybrasyl.Tests/Targeting.cs

Drop FormulaParserScansItemObjectForVariables: FormulaParser is
internal and the test never compiled (typeof on inaccessible type).
The two remaining FormulaVariable tests already cover the contract -
that 12 ItemObject properties carry the expected annotation.
